Rewrite XmlSchemaDemo
* All support classes are now nested classes inside XmlSchemaDemo
* Test two different implementation of the new `IXmlSchemaAssemblyCreator` interface
