+++ src/NodaTime.Demo/NodaTime.Demo.csproj
@@ -1,7 +1,6 @@
 ﻿<Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <OutputType>Exe</OutputType>
     <TargetFramework>netcoreapp3.1</TargetFramework>
     <Deterministic>True</Deterministic>
     <IsPackable>False</IsPackable>
@@ -13,30 +12,18 @@
 
   <ItemGroup>
     <ProjectReference Include="..\NodaTime\NodaTime.csproj" />
+    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="2.6.1" />
+    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.4.0" />
+    <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.0" PrivateAssets="All" IncludeAssets="runtime; build; native; contentfiles; analyzers" />
     <PackageReference Include="NUnit" Version="3.12.0" />
     <PackageReference Include="NUnit3TestAdapter" Version="3.16.1" />
-    <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.4.0" />
-  </ItemGroup>
-
-  <ItemGroup Condition="$(TargetFrameworkIdentifier) == '.NETFramework'">
-    <PackageReference Include="Microsoft.CSharp" Version="4.7.0" />
-    <PackageReference Include="Microsoft.NETFramework.ReferenceAssemblies" Version="1.0.0" PrivateAssets="All" IncludeAssets="runtime; build; native; contentfiles; analyzers" />
-  </ItemGroup>
-  <ItemGroup Condition="$(TargetFrameworkIdentifier) != '.NETFramework'">
-    <PackageReference Include="Microsoft.CodeAnalysis.CSharp" Version="2.6.1" />
     <PackageReference Include="XmlSchemaClassGenerator-beta" Version="2.0.349" />
   </ItemGroup>
 
   <ItemGroup>
-    <Compile Update="XmlSchemaDemo.Net*.cs">
+    <Compile Update="XmlSchemaDemo.*.cs">
       <DependentUpon>XmlSchemaDemo.cs</DependentUpon>
     </Compile>
-  </ItemGroup>
-  <ItemGroup Condition="$(TargetFrameworkIdentifier) == '.NETFramework'">
-    <Compile Remove="XmlSchemaDemo.NetCore.cs" />
-  </ItemGroup>
-  <ItemGroup Condition="$(TargetFrameworkIdentifier) != '.NETFramework'">
-    <Compile Remove="XmlSchemaDemo.NetFramework.cs" />
   </ItemGroup>
 
 </Project>
